// frontend/src/api/__generated__/models/DocListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Docsa API
 * Docsa의 백엔드 API 명세입니다.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RecentActivityDto } from './RecentActivityDto';
import {
    RecentActivityDtoFromJSON,
    RecentActivityDtoFromJSONTyped,
    RecentActivityDtoToJSON,
    RecentActivityDtoToJSONTyped,
} from './RecentActivityDto';

/**
 * 
 * @export
 * @interface DocListResponse
 */
export interface DocListResponse {
    /**
     * 
     * @type {number}
     * @memberof DocListResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DocListResponse
     */
    title?: string;
    /**
     * 
     * @type {Date}
     * @memberof DocListResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof DocListResponse
     */
    updatedAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof DocListResponse
     */
    preview?: string;
    /**
     * 
     * @type {RecentActivityDto}
     * @memberof DocListResponse
     */
    recent?: RecentActivityDto;
}

/**
 * Check if a given object implements the DocListResponse interface.
 */
export function instanceOfDocListResponse(value: object): value is DocListResponse {
    return true;
}

export function DocListResponseFromJSON(json: any): DocListResponse {
    return DocListResponseFromJSONTyped(json, false);
}

export function DocListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): DocListResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
        'preview': json['preview'] == null ? undefined : json['preview'],
        'recent': json['recent'] == null ? undefined : RecentActivityDtoFromJSON(json['recent']),
    };
}

export function DocListResponseToJSON(json: any): DocListResponse {
    return DocListResponseToJSONTyped(json, false);
}

export function DocListResponseToJSONTyped(value?: DocListResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
        'preview': value['preview'],
        'recent': RecentActivityDtoToJSON(value['recent']),
    };
}