// frontend/src/api/__generated__/models/PageDocListResponse.ts
export interface PageDocListResponse {
    /**
     * 
     * @type {number}
     * @memberof PageDocListResponse
     */
    totalPages?: number;
    /**
     * 
     * @type {number}
     * @memberof PageDocListResponse
     */
    totalElements?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof PageDocListResponse
     */
    pageable?: PageableObject;
    /**
     * 
     * @type {number}
     * @memberof PageDocListResponse
     */
    size?: number;
    /**
     * 
     * @type {Array<DocListResponse>}
     * @memberof PageDocListResponse
     */
    content?: Array<DocListResponse>;
    /**
     * 
     * @type {number}
     * @memberof PageDocListResponse
     */
    number?: number;
    /**
     * 
     * @type {SortObject}
     * @memberof PageDocListResponse
     */
    sort?: SortObject;
    /**
     * 
     * @type {number}
     * @memberof PageDocListResponse
     */
    numberOfElements?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PageDocListResponse
     */
    first?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageDocListResponse
     */
    last?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageDocListResponse
     */
    empty?: boolean;
}